Added code to not delete ds if ingest is running
Added code to not delete a datasource if ingest is running.

+++ Core/src/org/sleuthkit/autopsy/actions/DeleteDataSourceAction.java
@@ -19,11 +19,8 @@
 package org.sleuthkit.autopsy.actions;
 
 import java.awt.event.ActionEvent;
-import java.text.MessageFormat;
-import java.util.concurrent.ExecutionException;
 import java.util.logging.Level;
 import javax.swing.AbstractAction;
-import javax.swing.JOptionPane;
 import javax.swing.SwingWorker;
 import org.openide.DialogDescriptor;
 import org.openide.DialogDisplayer;
@@ -32,6 +29,7 @@ import org.openide.util.Lookup;
 import org.openide.util.NbBundle;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.casemodule.CaseActionException;
+import org.sleuthkit.autopsy.ingest.IngestManager;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.coreutils.MessageNotifyUtil;
 import org.sleuthkit.autopsy.keywordsearchservice.KeywordSearchService;
@@ -41,56 +39,64 @@ import org.sleuthkit.autopsy.keywordsearchservice.KeywordSearchServiceException;
  * Instances of this Action allow users to delete the specified data source.
  */
 public final class DeleteDataSourceAction extends AbstractAction {
+
     private static final Logger logger = Logger.getLogger(DeleteDataSourceAction.class.getName());
     private final Long selectedDataSource;
-    
+
     @NbBundle.Messages({"DeleteDataSourceAction.name.text=Delete Data Source"})
     public DeleteDataSourceAction(Long selectedDataSource) {
         super(Bundle.DeleteDataSourceAction_name_text());
         this.selectedDataSource = selectedDataSource;
     }
-    
+
     @NbBundle.Messages({"ErrorDeletingDataSource.name.text=Error Deleting Data Source",
-                        "DeleteDataSourceConfirmationDialog_message=Are you sure you want to delete the data source?",
-                        "DeleteDataSourceConfirmationDialog_title=Delete Data Source?"})
+        "ErrorDeletingDataSource.ingest.running=Cannot delete a Data Source when ingest is running.",
+        "DeleteDataSourceConfirmationDialog_message=Are you sure you want to delete the data source?",
+        "DeleteDataSourceConfirmationDialog_title=Delete Data Source?"})
     @Override
     public void actionPerformed(ActionEvent event) {
-        Object response = DialogDisplayer.getDefault().notify(new NotifyDescriptor(
-                Bundle.DeleteDataSourceConfirmationDialog_message(),
-                Bundle.DeleteDataSourceConfirmationDialog_title(),
-                NotifyDescriptor.YES_NO_OPTION,
-                NotifyDescriptor.WARNING_MESSAGE,
-                null,
-                NotifyDescriptor.NO_OPTION));
-        if (null != response && DialogDescriptor.YES_OPTION == response) {
+        if (!IngestManager.getInstance().isIngestRunning()) {
+            Object response = DialogDisplayer.getDefault().notify(new NotifyDescriptor(
+                    Bundle.DeleteDataSourceConfirmationDialog_message(),
+                    Bundle.DeleteDataSourceConfirmationDialog_title(),
+                    NotifyDescriptor.YES_NO_OPTION,
+                    NotifyDescriptor.WARNING_MESSAGE,
+                    null,
+                    NotifyDescriptor.NO_OPTION));
+            if (null != response && DialogDescriptor.YES_OPTION == response) {
 
-            new SwingWorker<Void, Void>() {
+                new SwingWorker<Void, Void>() {
 
-                @Override
-                protected Void doInBackground() throws Exception {
-                    try {
-                        Case.deleteDataSourceFromCurrentCase(selectedDataSource);
-                        deleteDataSource(selectedDataSource);
-                    } catch (CaseActionException | KeywordSearchServiceException ex) {
-                        logger.log(Level.WARNING, Bundle.ErrorDeletingDataSource_name_text(), ex);
-                        MessageNotifyUtil.Message.info(Bundle.ErrorDeletingDataSource_name_text());
+                    @Override
+                    protected Void doInBackground() throws Exception {
+                        try {
+                            Case.deleteDataSourceFromCurrentCase(selectedDataSource);
+                            deleteDataSource(selectedDataSource);
+                        } catch (CaseActionException | KeywordSearchServiceException ex) {
+                            logger.log(Level.WARNING, Bundle.ErrorDeletingDataSource_name_text(), ex);
+                            MessageNotifyUtil.Message.info(Bundle.ErrorDeletingDataSource_name_text());
+                        }
+                        return null;
                     }
-                    return null;
-                }
 
-                @Override
-                protected void done() {
-                }
-            }.execute();        
+                    @Override
+                    protected void done() {
+                    }
+                }.execute();
+            }
+        } else {
+            logger.log(Level.WARNING, Bundle.ErrorDeletingDataSource_ingest_running());
+            MessageNotifyUtil.Message.info(Bundle.ErrorDeletingDataSource_ingest_running());
+
         }
-    } 
-    
+    }
+
     private static void deleteDataSource(Long dataSourceId) throws KeywordSearchServiceException {
         try {
             KeywordSearchService kwsService = Lookup.getDefault().lookup(KeywordSearchService.class);
             kwsService.deleteDataSource(dataSourceId);
         } catch (KeywordSearchServiceException ex) {
             logger.log(Level.WARNING, "KWS Error", ex);
-        } 
+        }
     }
 }
